Tried fixing the errors but instead made things work...app isnt launching. gonna try changing podfile and reinstalling pods
