+++ frontend/.eslintignore
@@ -7,3 +7,4 @@ public/
 next.config.js
 vite.config.ts
 metro.config.js
+src/app/old_home.tsx
