+++ frontend/metro.config.js
@@ -1,20 +1,32 @@
-// metro.config.js
+// // metro.config.js
+// const { getDefaultConfig } = require("expo/metro-config");
+
+// module.exports = (() => {
+//   const config = getDefaultConfig(__dirname);
+
+//   const { transformer, resolver } = config;
+
+//   config.transformer = {
+//     ...transformer,
+//     babelTransformerPath: require.resolve("react-native-svg-transformer"),
+//   };
+//   config.resolver = {
+//     ...resolver,
+//     assetExts: resolver.assetExts.filter((ext) => ext !== "svg"),
+//     sourceExts: [...resolver.sourceExts, "svg"],
+//   };
+
+//   return config;
+// })();
+
+// frontend/metro.config.js
+
 const { getDefaultConfig } = require("expo/metro-config");
 
-module.exports = (() => {
-  const config = getDefaultConfig(__dirname);
+const defaultConfig = getDefaultConfig(__dirname);
 
-  const { transformer, resolver } = config;
+defaultConfig.transformer.babelTransformerPath = require.resolve("react-native-svg-transformer");
+defaultConfig.resolver.assetExts = defaultConfig.resolver.assetExts.filter((ext) => ext !== "svg");
+defaultConfig.resolver.sourceExts.push("svg");
 
-  config.transformer = {
-    ...transformer,
-    babelTransformerPath: require.resolve("react-native-svg-transformer"),
-  };
-  config.resolver = {
-    ...resolver,
-    assetExts: resolver.assetExts.filter((ext) => ext !== "svg"),
-    sourceExts: [...resolver.sourceExts, "svg"],
-  };
-
-  return config;
-})();
+module.exports = defaultConfig;
