+++ frontend/src/app/(tabs)/index.tsx
@@ -1,9 +1,11 @@
 import { StyleSheet, Text, View } from "react-native";
+import HomeIcon from "../../assets/home-icon.svg";
 
 export default function Home() {
   return (
     <View style={styles.container}>
       <Text style={styles.text}>Home</Text>
+      <HomeIcon />
     </View>
   );
 }
