+++ frontend/svg.d.ts
@@ -1,0 +1,7 @@
+// svg.d.ts
+declare module "*.svg" {
+  import React from "react";
+  import { SvgProps } from "react-native-svg";
+  const content: React.FC<SvgProps>;
+  export default content;
+}
